:truck: move: 파일 이동
- common 패키지에 이메일 양식 및 SendGrid 전송 클래스와 이메일 인증 전송 클래스 이동

- notification 패키지에는 이메일 알림 전송 관련 파일만 남김
