+++ src/main/java/com/project/cheerha/common/email/sender/EmailSender.java
@@ -1,4 +1,4 @@
-package com.project.cheerha.domain.notification.sender;
+package com.project.cheerha.common.email.sender;
 
 import java.io.IOException;
