+++ src/main/java/com/project/cheerha/domain/user/service/UserEmailVerificationService.java
@@ -6,7 +6,7 @@ import com.project.cheerha.common.exception.data.DataErrorCode;
 import com.project.cheerha.common.exception.data.NotFoundException;
 import com.project.cheerha.common.redis.EmailTokenService;
 import com.project.cheerha.domain.notification.sender.EmailSenderInit;
-import com.project.cheerha.domain.notification.sender.VerificationEmailSender;
+import com.project.cheerha.common.email.sender.VerificationEmailSender;
 import com.project.cheerha.domain.user.dto.response.ActivateNotificationResponseDto;
 import com.project.cheerha.domain.user.dto.response.VerifyPasswordResetTokenResponseDto;
 import com.project.cheerha.domain.user.dto.response.SendEmailVerificationResponseDto;
